feat(create-command): remove the create command
BREAKING CHANGE: The create command is superfluous and unnecessary

BREAKING CHANGE: The create command is superfluous and unnecessary
