--- src/executables/jb-create-cli-create.js
@@ -1,16 +1,0 @@
-#!/usr/bin/env node
-
-/* eslint-disable no-console */
-
-import createPackage from '../createPackage';
-
-const execute = async () => {
-  try {
-    await createPackage();
-  } catch (e) {
-    console.error('😞  Rut ro, an error occurred');
-    console.error(e);
-  }
-};
-
-execute();
